// src/proto/dapr/proto/runtime/v1/dapr.ts
export interface GetConfigurationRequest {
  /** Required. The name of configuration store. */
  storeName: string;
  /**
   * Optional. The key of the configuration item to fetch.
   * If set, only query for the specified configuration items.
   * Empty list means fetch all.
   */
  keys: string[];
  /** Optional. The metadata which will be sent to configuration store components. */
  metadata: { [key: string]: string };
}

export interface GetConfigurationRequest_MetadataEntry {
  key: string;
  value: string;
}
// ...
export interface SubscribeConfigurationRequest {
  /** The name of configuration store. */
  storeName: string;
  /**
   * Optional. The key of the configuration item to fetch.
   * If set, only query for the specified configuration items.
   * Empty list means fetch all.
   */
  keys: string[];
  /** The metadata which will be sent to configuration store components. */
  metadata: { [key: string]: string };
}

export interface SubscribeConfigurationRequest_MetadataEntry {
  key: string;
  value: string;
}

/** UnSubscribeConfigurationRequest is the message to stop watching the key-value configuration. */
export interface UnsubscribeConfigurationRequest {
  /** The name of configuration store. */
  storeName: string;
  /** The id to unsubscribe. */
  id: string;
}

export interface SubscribeConfigurationResponse {
  /** Subscribe id, used to stop subscription. */
  id: string;
  /** The list of items containing configuration values */
  items: { [key: string]: ConfigurationItem };
}
// ...
export const GetConfigurationRequest = {
  encode(message: GetConfigurationRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.storeName !== "") {
      writer.uint32(10).string(message.storeName);
    }
    for (const v of message.keys) {
      writer.uint32(18).string(v!);
    }
    Object.entries(message.metadata).forEach(([key, value]) => {
      GetConfigurationRequest_MetadataEntry.encode({ key: key as any, value }, writer.uint32(26).fork()).ldelim();
    });
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): GetConfigurationRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetConfigurationRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.storeName = reader.string();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.keys.push(reader.string());
          continue;
        case 3:
          if (tag !== 26) {
            break;
          }

          const entry3 = GetConfigurationRequest_MetadataEntry.decode(reader, reader.uint32());
          if (entry3.value !== undefined) {
            message.metadata[entry3.key] = entry3.value;
          }
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  // encodeTransform encodes a source of message objects.
  // Transform<GetConfigurationRequest, Uint8Array>
  async *encodeTransform(
    source:
      | AsyncIterable<GetConfigurationRequest | GetConfigurationRequest[]>
      | Iterable<GetConfigurationRequest | GetConfigurationRequest[]>,
  ): AsyncIterable<Uint8Array> {
    for await (const pkt of source) {
      if (globalThis.Array.isArray(pkt)) {
        for (const p of (pkt as any)) {
          yield* [GetConfigurationRequest.encode(p).finish()];
        }
      } else {
        yield* [GetConfigurationRequest.encode(pkt as any).finish()];
      }
    }
  },

  // decodeTransform decodes a source of encoded messages.
  // Transform<Uint8Array, GetConfigurationRequest>
  async *decodeTransform(
    source: AsyncIterable<Uint8Array | Uint8Array[]> | Iterable<Uint8Array | Uint8Array[]>,
  ): AsyncIterable<GetConfigurationRequest> {
    for await (const pkt of source) {
      if (globalThis.Array.isArray(pkt)) {
        for (const p of (pkt as any)) {
          yield* [GetConfigurationRequest.decode(p)];
        }
      } else {
        yield* [GetConfigurationRequest.decode(pkt as any)];
      }
    }
  },

  fromJSON(object: any): GetConfigurationRequest {
    return {
      storeName: isSet(object.storeName) ? globalThis.String(object.storeName) : "",
      keys: globalThis.Array.isArray(object?.keys) ? object.keys.map((e: any) => globalThis.String(e)) : [],
      metadata: isObject(object.metadata)
        ? Object.entries(object.metadata).reduce<{ [key: string]: string }>((acc, [key, value]) => {
          acc[key] = String(value);
          return acc;
        }, {})
        : {},
    };
  },

  toJSON(message: GetConfigurationRequest): unknown {
    const obj: any = {};
    if (message.storeName !== "") {
      obj.storeName = message.storeName;
    }
    if (message.keys?.length) {
      obj.keys = message.keys;
    }
    if (message.metadata) {
      const entries = Object.entries(message.metadata);
      if (entries.length > 0) {
        obj.metadata = {};
        entries.forEach(([k, v]) => {
          obj.metadata[k] = v;
        });
      }
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetConfigurationRequest>, I>>(base?: I): GetConfigurationRequest {
    return GetConfigurationRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetConfigurationRequest>, I>>(object: I): GetConfigurationRequest {
    const message = createBaseGetConfigurationRequest();
    message.storeName = object.storeName ?? "";
    message.keys = object.keys?.map((e) => e) || [];
    message.metadata = Object.entries(object.metadata ?? {}).reduce<{ [key: string]: string }>((acc, [key, value]) => {
      if (value !== undefined) {
        acc[key] = globalThis.String(value);
      }
      return acc;
    }, {});
    return message;
  },
};
// ...
export const SubscribeConfigurationRequest = {
  encode(message: SubscribeConfigurationRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.storeName !== "") {
      writer.uint32(10).string(message.storeName);
    }
    for (const v of message.keys) {
      writer.uint32(18).string(v!);
    }
    Object.entries(message.metadata).forEach(([key, value]) => {
      SubscribeConfigurationRequest_MetadataEntry.encode({ key: key as any, value }, writer.uint32(26).fork()).ldelim();
    });
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): SubscribeConfigurationRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSubscribeConfigurationRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.storeName = reader.string();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.keys.push(reader.string());
          continue;
        case 3:
          if (tag !== 26) {
            break;
          }

          const entry3 = SubscribeConfigurationRequest_MetadataEntry.decode(reader, reader.uint32());
          if (entry3.value !== undefined) {
            message.metadata[entry3.key] = entry3.value;
          }
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  // encodeTransform encodes a source of message objects.
  // Transform<SubscribeConfigurationRequest, Uint8Array>
  async *encodeTransform(
    source:
      | AsyncIterable<SubscribeConfigurationRequest | SubscribeConfigurationRequest[]>
      | Iterable<SubscribeConfigurationRequest | SubscribeConfigurationRequest[]>,
  ): AsyncIterable<Uint8Array> {
    for await (const pkt of source) {
      if (globalThis.Array.isArray(pkt)) {
        for (const p of (pkt as any)) {
          yield* [SubscribeConfigurationRequest.encode(p).finish()];
        }
      } else {
        yield* [SubscribeConfigurationRequest.encode(pkt as any).finish()];
      }
    }
  },

  // decodeTransform decodes a source of encoded messages.
  // Transform<Uint8Array, SubscribeConfigurationRequest>
  async *decodeTransform(
    source: AsyncIterable<Uint8Array | Uint8Array[]> | Iterable<Uint8Array | Uint8Array[]>,
  ): AsyncIterable<SubscribeConfigurationRequest> {
    for await (const pkt of source) {
      if (globalThis.Array.isArray(pkt)) {
        for (const p of (pkt as any)) {
          yield* [SubscribeConfigurationRequest.decode(p)];
        }
      } else {
        yield* [SubscribeConfigurationRequest.decode(pkt as any)];
      }
    }
  },

  fromJSON(object: any): SubscribeConfigurationRequest {
    return {
      storeName: isSet(object.storeName) ? globalThis.String(object.storeName) : "",
      keys: globalThis.Array.isArray(object?.keys) ? object.keys.map((e: any) => globalThis.String(e)) : [],
      metadata: isObject(object.metadata)
        ? Object.entries(object.metadata).reduce<{ [key: string]: string }>((acc, [key, value]) => {
          acc[key] = String(value);
          return acc;
        }, {})
        : {},
    };
  },

  toJSON(message: SubscribeConfigurationRequest): unknown {
    const obj: any = {};
    if (message.storeName !== "") {
      obj.storeName = message.storeName;
    }
    if (message.keys?.length) {
      obj.keys = message.keys;
    }
    if (message.metadata) {
      const entries = Object.entries(message.metadata);
      if (entries.length > 0) {
        obj.metadata = {};
        entries.forEach(([k, v]) => {
          obj.metadata[k] = v;
        });
      }
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SubscribeConfigurationRequest>, I>>(base?: I): SubscribeConfigurationRequest {
    return SubscribeConfigurationRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SubscribeConfigurationRequest>, I>>(
    object: I,
  ): SubscribeConfigurationRequest {
    const message = createBaseSubscribeConfigurationRequest();
    message.storeName = object.storeName ?? "";
    message.keys = object.keys?.map((e) => e) || [];
    message.metadata = Object.entries(object.metadata ?? {}).reduce<{ [key: string]: string }>((acc, [key, value]) => {
      if (value !== undefined) {
        acc[key] = globalThis.String(value);
      }
      return acc;
    }, {});
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends Long ? string | number | Long : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };